refactor: update name of movement package
- from movement to commands
