+++ internal/app/commands/mocks/rotatable.go
