+++ internal/app/commands/movable.go
@@ -1,4 +1,4 @@
-package movement
+package commands
 
 import (
 	"github.com/dosovma/otus_arch/pkg"
