+++ internal/app/commands/rotatable_test.go
@@ -1,13 +1,13 @@
-package movement_test
+package commands_test
 
 import (
 	"errors"
 	"testing"
 
-	"github.com/dosovma/otus_arch/internal/app/movement"
+	"github.com/dosovma/otus_arch/internal/app/commands"
 	"github.com/golang/mock/gomock"
 
-	"github.com/dosovma/otus_arch/internal/app/movement/mocks"
+	"github.com/dosovma/otus_arch/internal/app/commands/mocks"
 )
 
 func TestRotate_Execute(t *testing.T) {
@@ -73,7 +73,7 @@ func TestRotate_Execute(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			mockRotate := mocks.NewMockRotatable(gomock.NewController(t))
-			rm := movement.Rotate{
+			rm := commands.Rotate{
 				Rotate: mockRotate,
 			}
 			tt.mockCalls(mockRotate)
